tests: dead code warnings are fine when tests are disabled
